AIRAVATA-2339 Attempting to fix data staging with individual account
In ProcessContext I removed use of isUseUserCRPref() in methods related to
getting gateway storage information. Whether the current experiment is submitted
with user's compute resource preferences or not should not affect gateway
storage information. How to specify user's storage preferences is not yet
implemented and must be implemented with something other than isUseUserCRPref().

In SCPDataStageTask, I changed the construction of the sshSession for the SCP
connection to the compute resource to use
getComputerResourceSSHKeyAuthentication() instead of
getStorageSSHKeyAuthentication(), which would not be correct, if I understand
the code correctly.