AIRAVATA-2347 Removing use of gatewayResource, GATEWAY_WORKER stuff
I don't think the GATEWAY_WORKER stuff is needed any longer. Also, the
ExperimentCatalog is instantiated in the RegistryServerHandler as a singleton
with reference to a gatewayId, so it has the incorrect gatewayId for this request and can't be
trusted.  This shows up here where the gatewayResource actually references the
default gateway from the airavata-server.properties file instead of the gateway
in the request.