AIRAVATA-2821 Setting status id in Process, Job Repositories
Also changed the order of implicit field initialization in save methods
so that model fields are updated first, then models are mapped to entity
classes, then entity fields (mostly primary key fields) are updated
last. This keeps the model fields up to date with what is persisted and
fixes the problem of creating a duplicate status if an experiment is
saved and then updated.