AIRAVATA-2872 Allow setting resourceSpecificCredStoreToken to null
Changed the "detach state" setting in OpenJPA to "all" which changes how
null values are merged. Without "all", null values are treated as simply
unspecified and are not merged in. With "all", null values are treated
as specified values that are merged in.

One consequence is that any field on the JPA entities that isn't in the
Thrift models OpenJPA will attempt to set it to null. For this reason,
@ManyToOne references, which aren't in the Thrift models, are configured
so that the @JoinColumn is not nullable and not updateable, which is
true regardless but is necessary now with the change in "detach state"
setting.

I also set up orphanRemoval on the GroupResourceProfileEntity to test
that that would still work as well.