AIRAVATA-2984 Generate webpack files in STATIC_ROOT dir
Generate the webpack-stats.json file and the webpack generated files directly
into STATIC_ROOT for a production deploy. This way the paths to the
generated files in webpack-stats.json will be to the locations under
STATIC_ROOT. Otherwise the webpack-stats.json locations are pointing to
the source directory but these get cleared out during a build/deploy.