AIRAVATA-3403 Fix file permissions of large file uploads
Uploads of Wagtail media over 2.5 MB are uploaded to the tmp dir
and then moved into the media folder. Their permissions in tmp dir
before this setting were os-dependent but in our production env
it was user-readable only (0x600). This fix ensures these files
are world-readable and so are accessible to Apache.