Added module commands to template to allow for custom HTCondor configuration attributes
Since HTCondor has a lot of custom attributes depending on the pool or application at hand, there needs to be a way to allow for custom attributes to be added. Therefore, we've retrofitted the "module commands" array for this task, even though it's not the best. This will most likely cause issues in the future if an application is menat to work on multiple types of job management workflows such as SLURM where module commands are used to load specific libraries and programs to run an  experiment.